Refactor repository and branch page components for improved structure
- Consolidated repository and branch page logic into dedicated components for better maintainability and readability.
- Introduced loading states for enhanced user experience during data fetching.
- Updated routing to streamline redirection logic and improve performance.
- Removed outdated code and unnecessary complexity from previous implementations.
