+++ components/header.tsx
@@ -1,9 +1,6 @@
-import { SignedIn, SignedOut, SignInButton, UserButton } from '@clerk/nextjs'
-import { Key } from 'lucide-react'
 import Image from 'next/image'
 import Link from 'next/link'
-import { ThemeToggle } from '@/components/theme-toggle'
-import { Button } from '@/components/ui/button'
+import { HeaderButtons } from '@/components/header-buttons'
 
 export function Header() {
   return (
@@ -19,25 +16,7 @@ export function Header() {
             width={100}
           />
         </Link>
-        <div className="flex items-center gap-4">
-          <SignedIn>
-            <Link href="/token">
-              <Button className="gap-2" size="sm" variant="ghost">
-                <Key className="h-4 w-4" />
-                <span className="hidden sm:inline">API Key</span>
-              </Button>
-            </Link>
-            <UserButton />
-          </SignedIn>
-          <SignedOut>
-            <SignInButton mode="modal">
-              <Button size="sm" type="button" variant="link">
-                Sign In
-              </Button>
-            </SignInButton>
-          </SignedOut>
-          <ThemeToggle />
-        </div>
+        <HeaderButtons />
       </div>
     </header>
   )
